+++ R/new_model_server.R
@@ -209,19 +209,31 @@ new_model_server <- function(session, input, output, resources ){
   output$mapNAUI <- renderUI({
     if ( input$platformInput != "NONMEM") {
       return(
-        HTML_info("Variable mapping is not available for the selected software platform")
+        message_box(
+          text = "Variable mapping is not available for the selected software platform",
+          icon = "info-circle-fill",
+          theme = "info"
+        )
       )
     } else {
 
       if ( inherits(try(dataFileReactive(), silent = TRUE), "try-error") ){
         return(
-          HTML_info("No data file was selected")
+          message_box(
+            text = "No data file was selected",
+            icon = "info-circle-fill",
+            theme = "info"
+          )
         )
       }
 
       if ( inherits(dataFile(), "try-error") ){
         return(
-          HTML_info("Invalid data file")
+          message_box(
+            text = "Invalid data file",
+            icon = "x-circle-fill",
+            theme = "danger"
+          )
         )
       }
 
@@ -1289,6 +1301,9 @@ new_model_server <- function(session, input, output, resources ){
   })
 
   output$effectDriverUI <- renderUI({
+
+    req( input$pkInput, input$pdInput )
+
     max <- value <- 1
     if ( input$pkInput == "pk" ){
       max <- as.numeric(input$pkCMTInput)
@@ -1381,7 +1396,11 @@ new_model_server <- function(session, input, output, resources ){
       return(
         fluidRow(
           col_12(
-            HTML_info("No model structure defined")
+            message_box(
+              text = "No model structure defined",
+              icon = "info-circle-fill",
+              theme = "info"
+            )
           )
         )
       )
@@ -1389,9 +1408,27 @@ new_model_server <- function(session, input, output, resources ){
 
   })
 
+  scales <- reactive({
+
+    req( input$pkInput, input$pdInput, input$nOTParmInput )
+
+    req( input$pkInput != "none" | input$pdInput != "none" )
+
+    scales <- c("Linear", "Log", "Logit")
+
+    if ( any( input$pdInput %in% c("logistic", "ordcat") ) ){
+      scales <- "Linear"
+    }
+
+    scales
+
+  })
+
+  ### Parameter matrix
+
   parameterDF <- reactive({
 
-    req( input$pkInput, input$pdInput, input$nOTParmInput, input$new_menu == "Parameters" )
+    req( input$pkInput, input$pdInput, scales(), input$nOTParmInput, input$new_menu == "Parameters" )
 
     req( input$pkInput != "none" | input$pdInput != "none" )
 
@@ -1412,6 +1449,7 @@ new_model_server <- function(session, input, output, resources ){
         Max = character(),
         Fixed = character(),
         Variability = character(),
+        Scale = character(),
         stringsAsFactors = FALSE
       )
 
@@ -1447,6 +1485,7 @@ new_model_server <- function(session, input, output, resources ){
         Max = parm_info$MAX,
         Fixed = "No",
         Variability = parm_info$VAR,
+        Scale = "Linear",
         stringsAsFactors = FALSE
       )
 
@@ -1472,6 +1511,7 @@ new_model_server <- function(session, input, output, resources ){
                 Max = "+INF",
                 Fixed = "No",
                 Variability = "exp",
+                Scale = "Linear",
                 stringsAsFactors = FALSE
               )
             )
@@ -1495,6 +1535,7 @@ new_model_server <- function(session, input, output, resources ){
                 Max = "+INF",
                 Fixed = "No",
                 Variability = "exp",
+                Scale = "Linear",
                 stringsAsFactors = FALSE
               )
             )
@@ -1516,6 +1557,7 @@ new_model_server <- function(session, input, output, resources ){
               Max = "+INF",
               Fixed = "No",
               Variability = "none",
+              Scale = "Linear",
               stringsAsFactors = FALSE
             )
           )
@@ -1534,6 +1576,7 @@ new_model_server <- function(session, input, output, resources ){
               Max = "1",
               Fixed = "No",
               Variability = "none",
+              Scale = "Linear",
               stringsAsFactors = FALSE
             )
           )
@@ -1602,6 +1645,7 @@ new_model_server <- function(session, input, output, resources ){
         Max = "+INF",
         Fixed = "No",
         Variability = ifelse(grepl("ALAG", PKparms), "none", "exp"),
+        Scale = "Linear",
         stringsAsFactors = FALSE
       ) %>%
         dplyr::mutate(
@@ -1623,6 +1667,7 @@ new_model_server <- function(session, input, output, resources ){
               Max = "1",
               Fixed = "No",
               Variability = "none",
+              Scale = "Linear",
               stringsAsFactors = FALSE
             )
           )
@@ -1812,6 +1857,7 @@ new_model_server <- function(session, input, output, resources ){
         Max = character(),
         Fixed = character(),
         Variability = character(),
+        Scale = character(),
         stringsAsFactors = FALSE
       )
     } else
@@ -1827,6 +1873,7 @@ new_model_server <- function(session, input, output, resources ){
         Max = parm_info$MAX,
         Fixed = "No",
         Variability = parm_info$VAR,
+        Scale = "Linear",
         stringsAsFactors = FALSE
       )
     }
@@ -1846,6 +1893,7 @@ new_model_server <- function(session, input, output, resources ){
         Max = character(),
         Fixed = character(),
         Variability = character(),
+        Scale = character(),
         stringsAsFactors = FALSE
       )
     } else
@@ -1876,6 +1924,7 @@ new_model_server <- function(session, input, output, resources ){
         Max = parm_info$MAX,
         Fixed = "No",
         Variability = parm_info$VAR,
+        Scale = "Linear",
         stringsAsFactors = FALSE
       )
 
@@ -1905,7 +1954,8 @@ new_model_server <- function(session, input, output, resources ){
             .data$Variability == "logit" ~ "3",
             TRUE ~ NA_character_
           )
-        )
+        ),
+        Scale = factor("Linear", levels = scales(), ordered = TRUE),
       )
 
     ### Check content of input$parameterTable and preserve custom inputs
@@ -1916,6 +1966,7 @@ new_model_server <- function(session, input, output, resources ){
         DF <- oDF
       }
       if ( !identical(DF, oDF)) {
+
         mDF <- merge(
           cbind(
             DF,
@@ -1956,13 +2007,13 @@ new_model_server <- function(session, input, output, resources ){
 
   output$parameterTable <- renderRHandsontable({
 
-    req( parameterDF() )
+    req( parameterDF(), scales() )
 
     DF <- parameterDF()
 
     # Convert select variable to character (remove the glue class)
     DF <- DF %>%
-      mutate(
+      dplyr::mutate(
         Type = as.character(Type),
         SourceParam = as.character(SourceParam),
         Parameter = as.character(Parameter)
@@ -1974,7 +2025,7 @@ new_model_server <- function(session, input, output, resources ){
       contextMenu = FALSE,
       manualColumnMove = FALSE,
       manualRowMove = TRUE,
-      width = ifelse( input$platformInput == "NONMEM", 690, 640),
+      width = ifelse( input$platformInput == "NONMEM", 690, 640) + 60,
       height = max(200, (nrow(DF) + 1)*25 + 10)  # 25 px per row + 10 for potential scroll bar
     ) %>%
       hot_table(contextMenu = FALSE) %>%
@@ -1986,6 +2037,12 @@ new_model_server <- function(session, input, output, resources ){
       hot_col(col = "Min", colWidths = 50) %>%
       hot_col(col = "Initial", colWidths = 50) %>%
       hot_col(col = "Max", colWidths = 50) %>%
+      hot_col(
+        col = "Scale",
+        type = "dropdown",
+        source = scales(),
+        colWidths = 60
+      ) %>%
       hot_col(col = "Variability", colWidths = 0.1) %>% # Hide Variability
       hot_col(
         col = ifelse(
@@ -2038,13 +2095,11 @@ new_model_server <- function(session, input, output, resources ){
 
   outputOptions(output, "parameterTable", suspendWhenHidden = FALSE)
 
-  output$parameterTableUI <- renderUI( {
+  output$parameterTableUI <- renderUI({
     fluidRow(
-      col_12(
-        rHandsontableOutput("parameterTable")
-        )
-      )
-  } )
+      col_12( rHandsontableOutput("parameterTable") )
+    )
+  })
 
   parameterTable_content <- reactive({
     if ( is.null(input$parameterTable) | length(input$parameterTable$data) == 0) {
@@ -2056,7 +2111,7 @@ new_model_server <- function(session, input, output, resources ){
 
   output$parameterUI <- renderUI({
 
-    req( input$pkInput, input$pdInput)
+    req( input$pkInput, input$pdInput )
 
     if ( input$platformInput == 'NONMEM' & input$pdInput != 'logistic' & input$pdInput != 'ordcat' ){
       muBtn <- col_3(
@@ -2104,6 +2159,8 @@ new_model_server <- function(session, input, output, resources ){
       nPD <- NULL
     }
 
+    #
+
     fluidRow(
       col_12(
         fluidRow(
@@ -2127,7 +2184,6 @@ new_model_server <- function(session, input, output, resources ){
         ),
         fluidRow(
           col_12(
-            strong("Parameters"),
             uiOutput("parameterTableUI")
           )
         )
@@ -2138,16 +2194,17 @@ new_model_server <- function(session, input, output, resources ){
 
   outputOptions(output, "parameterUI", suspendWhenHidden = FALSE)
 
-  output$duplicateParmWarningUI <- renderUI({
+  parameterWarnings <- reactive({
+
     if ( is.null(input$parameterTable) | length(input$parameterTable$data) == 0 ){
       NULL
     } else {
-      parms <- hot_to_r(input$parameterTable)$Parameter
-      dupParms <- unique(parms[duplicated(parms)])
-      if ( length(dupParms) == 0 ){
-        NULL
-      } else {
-        HTML_info(
+      parms <- hot_to_r(input$parameterTable)
+
+      # Check for duplicate parameters
+      dupParms <- unique( parms[duplicated(parms$Parameter), "Parameter" ] )
+      if ( length(dupParms) > 0 ){
+        return(
           glue::glue(
             paste(
               "The parameter table includes duplicates ({vars}). Edit the parameter",
@@ -2156,19 +2213,132 @@ new_model_server <- function(session, input, output, resources ){
             vars = paste(dupParms, collapse = ", ")
           )
         )
-
       }
+
+      # Check for character initials
+      if ( any( is.na( suppressWarnings( as.numeric(parms$Initial) ) ) ) ) {
+        return(
+          "Initial values must be numeric."
+        )
+      }
+
+      if ( any(tolower(parms$Min) == '+inf') ) {
+        return(
+          "Minimum bounds cannot be set to +INF."
+        )
+      }
+
+      if ( any(tolower(parms$Max) == '-inf') ) {
+        return(
+          "Maximum bounds cannot be set to -INF."
+        )
+      }
+
+      mins <- parms$Min[ tolower(parms$Min) != "-inf" ]
+      maxs <- parms$Max[ tolower(parms$Max) != "+inf" ]
+
+      if ( any( is.na( suppressWarnings( as.numeric(mins) ) ) ) ) {
+        return(
+          "Lower boundaries must be numeric or -INF."
+        )
+      }
+      if ( any( is.na( suppressWarnings( as.numeric(maxs) ) ) ) ){
+        return(
+          "Upper boundaries must be numeric or +INF."
+        )
+      }
+
+      parms <- parms %>%
+        dplyr::mutate(
+          Min = as.numeric( sub('INF', 'Inf', Min) ),
+          Initial = as.numeric( sub('INF', 'Inf', Initial) ),
+          Max = as.numeric( sub('INF', 'Inf', Max) ),
+          minmax_check = Min >= Max,
+          initial_check = Min >= Initial | Initial >= Max,
+          logit_check = ifelse(
+            Scale == "Logit",
+            Min == -Inf | Max == Inf,
+            FALSE
+          )
+        )
+
+      if ( any(parms$minmax_check) ){
+        return(
+          "Lower boundaries must be lower than upper boundaries."
+        )
+      }
+      if ( any(parms$initial_check) ){
+        return(
+          "Initial values must be strictly between the lower and upper boundaries."
+        )
+      }
+      if ( any(parms$logit_check) ){
+        return(
+          "Boundaries must be finite for parameters defined on the logit scale."
+        )
+      }
+    }
+  })
+
+  output$parmInfoUI <- renderUI({
+    if ( is.null(input$parameterTable) | length(input$parameterTable$data) == 0 ){
+      NULL
+    } else {
+      parms <- hot_to_r(input$parameterTable)
+
+      text <- "All initial values and boundaries must be provided on the linear scale. Transformations to the log or logit scale will be automatically applied, whenever applicable."
+
+      if ( any( parms$Scale == 'Logit' ) & input$platformInput == 'NONMEM' ){
+        text <- c(
+          text,
+          "If required for the calculation of magnitude of variability, the inv_logit function is defined as follows:",
+          "inv_logit <- function(x, min = 0, max = 1){ min + (max-min)/(1+exp(-x)) }"
+        )
+      }
+
+      if ( any( parms$Scale != 'Linear' ) ){
+        message_box(
+          text = text,
+          icon = "info-circle-fill",
+          theme = "info"
+        )
+      } else {
+        NULL
+      }
+
+    }
+  })
+
+  output$duplicateParmWarningUI <- renderUI({
+
+    if ( !isTruthy(parameterWarnings()) ) {
+      NULL
+    } else {
+      message_box(
+        text = parameterWarnings(),
+        icon = "cone-striped",
+        theme = "warning"
+      )
     }
   })
 
   output$importParameterUI <- renderUI({
 
-    if ( input$platformInput != 'mrgsolve' ){
+    if ( input$platformInput != 'mrgsolve' | notTruthy(input$pkInput, input$pdInput) ){
       return(NULL)
     }
 
     fluidRow(
-      col_6(
+      col_3(
+        radioButtons(
+          inputId = "posthocInput",
+          label = "Include POSTHOC estimates",
+          choices = c("Yes", "No"),
+          selected = "No",
+          inline = TRUE
+        )
+      ),
+      col_3(
         radioButtons(
           inputId = "nmextInput",
           label = "Import NONMEM estimates",
@@ -2217,14 +2387,18 @@ new_model_server <- function(session, input, output, resources ){
     if ( notTruthy(input$pkInput, input$pdInput) ){
       fluidRow(
         col_12(
-          HTML_info("No model structure defined")
+          message_box(
+            text = "No model structure defined",
+            icon = "info-circle-fill",
+            theme = "info"
+          )
         )
       )
     }
 
   })
 
-  output$varianceTable <- renderRHandsontable({
+  varianceDF <- reactive({
 
     if ( input$platformInput == "Berkeley Madonna") {
       return(NULL)
@@ -2256,6 +2430,50 @@ new_model_server <- function(session, input, output, resources ){
     DF <- parameterTable %>%
       dplyr::select(.data$Variability, .data$Parameter)
     row.names(DF) <- DF$Parameter
+
+    ### Check content of input$varianceTable and preserve custom inputs
+    if ( length(isolate(input$varianceTable)) > 0 )
+    {
+      oDF <- hot_to_r(isolate(input$varianceTable))
+      if ( nrow(DF) == 0 ){
+        DF <- oDF
+      }
+      if ( !identical(DF, oDF)) {
+
+        mDF <- merge(
+          cbind(
+            DF,
+            data.frame("_SORT_" = 1:nrow(DF))
+          ),
+          oDF,
+          by = "Parameter",
+          all.x = TRUE
+        )
+        mDF <- mDF[order(mDF[, "X_SORT_"]), ]
+
+        for (col in names(oDF)[-2] ){
+          DF[, col] <- ifelse(
+            is.na(mDF[,paste(col, "y", sep = ".")]),
+            mDF[, paste(col, "x", sep = ".")],
+            mDF[, paste(col, "y", sep = ".")]
+          )
+          if ( is.factor(oDF[, col]) ){
+            # ifelse coerces factors to integers, must reset to factor
+            DF[, col] <- factor(levels(oDF[, col])[DF[, col]], levels = levels(oDF[, col]), ordered = TRUE)
+          }
+        }
+      }
+    }
+
+    DF
+
+  })
+
+  output$varianceTable <- renderRHandsontable({
+
+    req( varianceDF )
+
+    DF <- varianceDF()
 
     rhandsontable(
       data = DF,
@@ -2425,7 +2643,7 @@ new_model_server <- function(session, input, output, resources ){
   # Process covariance matrix and check if there are errors
   covarianceBlocks <- reactive({
 
-    chk1 <- chk2 <- chk3 <- chk4 <- TRUE
+    chk1 <- chk2 <- chk3 <- chk4 <- chk5 <- TRUE
     blocks <- NULL
 
     if ( isTruthy(input$covarianceTable) ){
@@ -2439,8 +2657,23 @@ new_model_server <- function(session, input, output, resources ){
         )
       )
 
-      ## Check 1: detect  if variance is set to 0 on a parameter with variability
       covarianceTable <- as.matrix(hot_to_r(input$covarianceTable))
+
+      if ( all( covarianceTable == 0, na.rm = TRUE ) ){
+        return(
+          list(
+            chk1 = chk1, chk2 = chk2, chk3 = chk3, chk4 = chk4,
+            blocks = list(
+              list(
+                omega = covarianceTable,
+                type = 'diagonal'
+              )
+            )
+          )
+        )
+      }
+
+      ## Check 1: detect  if variance is set to 0 on a parameter with variability
       n <- nrow(covarianceTable)
       diag(covarianceTable)[is.na(diag(covarianceTable))] <- 0
 
@@ -2558,34 +2791,63 @@ new_model_server <- function(session, input, output, resources ){
 
         }
       }
+
+      ## Check 5: check variability model
+      chk5 <- hot_to_r(input$parameterTable) %>%
+        dplyr::left_join(
+          hot_to_r(input$varianceTable),
+          by = 'Parameter'
+        ) %>%
+        dplyr::mutate(
+          invalid = case_when(
+            Scale == 'Log' & Variability.y == 'Logit' ~ TRUE,
+            Scale == 'Logit' & Variability.y != 'Logit' ~ TRUE,
+            TRUE ~ FALSE
+          )
+        ) %>%
+        filter( invalid == TRUE ) %>%
+        pull( Parameter )
+
+
     }
 
-    list(chk1 = chk1, chk2 = chk2, chk3 = chk3, chk4 = chk4, blocks = blocks)
+    list(chk1 = chk1, chk2 = chk2, chk3 = chk3, chk4 = chk4, chk5 = chk5, blocks = blocks)
 
   })
 
   ## Dynamic UI for misspecified covariance matrix
   output$covarianceWarningUI <- renderUI({
 
-    if ( all(c(covarianceBlocks()$chk1, covarianceBlocks()$chk2, covarianceBlocks()$chk3, covarianceBlocks()$chk4)) ){
+    if ( all(c(covarianceBlocks()$chk1, covarianceBlocks()$chk2, covarianceBlocks()$chk3, covarianceBlocks()$chk4, length(covarianceBlocks()$chk5) == 0 ) ) ){
       NULL
     } else {
       tagList(
         if ( !covarianceBlocks()$chk1 ) {
-          HTML_info(
-            "Variance cannot be set to 0 for a parameter with estimated variance"
+          message_box(
+            text = "Variance cannot be set to 0 for a parameter with estimated variance",
+            icon = "cone-striped",
+            theme = "warning"
           )
         },
-        if ( !covarianceBlocks()$chk1 ) {
-          p()
-        },
         if ( any(!c(covarianceBlocks()$chk2, covarianceBlocks()$chk3, covarianceBlocks()$chk4)) ) {
-          HTML_info(
-            paste(
+          message_box(
+            text = paste(
               "The covariance matrix must be constructed as a series of diagonal,",
               "band, or full block matrices. Correlation will be ignored in",
               "the model."
-            )
+            ),
+            icon = "cone-striped",
+            theme = "warning"
+          )
+        },
+        if ( length( covarianceBlocks()$chk5 ) > 0 ) {
+          message_box(
+            text = paste(
+              "Invalid variability selection for:",
+              paste(covarianceBlocks()$chk5, collapse = ", ")
+            ),
+            icon = "cone-striped",
+            theme = "warning"
           )
         }
       )
@@ -2599,13 +2861,21 @@ new_model_server <- function(session, input, output, resources ){
 
     if ( input$platformInput == "Berkeley Madonna") {
       return(
-        HTML_info("Variance-covariance settings are not available for the selected software platform")
+        message_box(
+          text = "Variance-covariance settings are not available for the selected software platform",
+          icon = "info-circle-fill",
+          theme = "info"
+        )
       )
     }
 
     if ( notTruthy(input$parameterTable) ){
       return(
-        HTML_info("No parameters defined")
+        message_box(
+          text = "No parameters defined",
+          icon = "info-circle-fill",
+          theme = "info"
+        )
       )
     }
 
@@ -2638,14 +2908,20 @@ new_model_server <- function(session, input, output, resources ){
 
     if ( input$platformInput == "Berkeley Madonna" ) {
       return(
-        HTML_info("Residual variability cannot be defined for the selected software platform")
+        message_box(
+          text = "Residual variability cannot be defined for the selected software platform",
+          icon = "info-circle-fill",
+          theme = "info"
+        )
       )
     }
 
     if ( notTruthy(input$pkInput, input$pdInput) ){
       fluidRow(
-        col_12(
-          HTML_info("No model structure defined")
+        message_box(
+          text = "No model structure defined",
+          icon = "info-circle-fill",
+          theme = "info"
         )
       )
     }
@@ -2920,7 +3196,11 @@ new_model_server <- function(session, input, output, resources ){
     if ( rvCheck()$isOK ){
       NULL
     } else {
-      HTML_info("Variance cannot be set to 0 for an estimated RV variance")
+      message_box(
+        text = "Variance cannot be set to 0 for an estimated RV variance",
+        icon = "cone-striped",
+        theme = "warning"
+      )
     }
   })
 
@@ -3123,8 +3403,10 @@ new_model_server <- function(session, input, output, resources ){
     if ( notTruthy(input$pkInput, input$pdInput) ){
       return(
         fluidRow(
-          col_12(
-            HTML_info("No model structure defined")
+          message_box(
+            text = "No model structure defined",
+            icon = "info-circle-fill",
+            theme = "info"
           )
         )
       )
@@ -3195,8 +3477,10 @@ new_model_server <- function(session, input, output, resources ){
     if ( notTruthy(input$pkInput, input$pdInput) ){
       return(
         fluidRow(
-          col_12(
-            HTML_info("No model structure defined")
+          message_box(
+            text = "No model structure defined",
+            icon = "info-circle-fill",
+            theme = "info"
           )
         )
       )
@@ -3400,13 +3684,17 @@ new_model_server <- function(session, input, output, resources ){
         length(covarianceTable) > 0
       )
 
-      if ( length(parameterTable$Parameter) != length(unique(parameterTable$Parameter)) ){
-        return("Duplicates in parameter tables prevents the code generation.")
+      if ( isTruthy(parameterWarnings()) ){
+        return("Invalid parameter definition prevents the code generation.")
       } else if (
         nrow(parameterTable) != nrow(varianceTable) |
         nrow(varianceTable) != nrow(covarianceTable)
       ) {
-        return("Inconsistent dimension of parameter and variance/covariance tables.")
+        return("Inconsistent dimension of parameter and variance/covariance tables prevents the code generation.")
+      } else if (
+        !all(c(covarianceBlocks()$chk1, covarianceBlocks()$chk2, covarianceBlocks()$chk3, covarianceBlocks()$chk4, length(covarianceBlocks()$chk5) == 0 ) )
+      ){
+        return("Invalid variability selection prevents the code generation")
       } else {
 
         get_code(