+++ R/zzz.R
@@ -6,7 +6,8 @@
     c("parm_lib", "pdForm_lib", "labelunit_lib", "rv_lib", "scaling",
       "template_nonmem", "template_mrgsolve", "template_bm", "model_lib",
       "Action", "Center", "Covariate", "Flags", "Function", "Functional form",
-      "Initial", "SourceParam", "Parameter", 'Stage', 'Step', "Type", "combo", "name"
+      "Initial", "SourceParam", "Parameter", "Stage", "Step", "Type", "combo", "name",
+      "Max", "Min", "Scale", "invalid"
       )
   )